+++ changelog.md
@@ -18,6 +18,7 @@ This version is not yet released. If you are reading this on the website, then t
   - This should not affect any language semantics
 - Improve pattern matching error messages
 - Optimize the "root" pattern `ⁿ%:1`
+- Add an `-e`/`--experimental` flag to the `uiua eval` command to enable experimental features
 ### Website
 - Add a new pad setting to show line values to the right of the code